+++ pootle/apps/pootle_app/forms.py
@@ -57,7 +57,7 @@ class ProjectForm(forms.ModelForm):
     class Meta(object):
         model = Project
         fields = ('id', 'code', 'fullname', 'checkstyle',
-                  'filetypes', 'source_language', 'ignoredfiles',
+                  'filetypes', 'source_language',
                   'report_email', 'screenshot_search_prefix', 'disabled',)
 
     def __init__(self, *args, **kwargs):