+++ pootle/apps/pootle_app/management/commands/dump.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
@@ -33,8 +34,7 @@ DUMPED = {
     'Suggestion': ('target_f', 'user_id'),
     'Language': ('code', 'fullname', 'pootle_path'),
     'Project': ('code', 'fullname', 'checkstyle',
-                'source_language', 'ignoredfiles',
-                'screenshot_search_prefix', 'disabled')
+                'source_language', 'screenshot_search_prefix', 'disabled')
 }